refactor: move InfoItem into components folder
