+++ src/components/InfoItem.tsx
